[REFACTOR] split game.hpp into more files
3dgui.hpp
fpsstate.hpp
gamemode.hpp
weapon.hpp
